requirements.txt
requests
pycryptodome
ddddocr